rename `OnyxSeparator` to `OnyxNavSeparator`
