+++ packages/sit-onyx/src/components/OnyxNavSeparator/OnyxNavSeparator.ct.tsx
@@ -1,13 +1,13 @@
 import { expect, test } from "../../playwright/a11y";
 import { executeMatrixScreenshotTest } from "../../playwright/screenshots";
-import OnyxSeparator from "./OnyxSeparator.vue";
+import OnyxNavSeparator from "./OnyxNavSeparator.vue";
 
 test.describe("Screenshot tests", () => {
   executeMatrixScreenshotTest({
-    name: "Separator",
+    name: "Nav separator",
     columns: ["default"],
     rows: ["default"],
-    component: () => <OnyxSeparator />,
+    component: () => <OnyxNavSeparator />,
     beforeScreenshot: async (component) => {
       await expect(component.getByRole("separator")).toHaveAttribute(
         "aria-orientation",
