+++ packages/sit-onyx/src/components/OnyxNavSeparator/OnyxNavSeparator.stories.ts
@@ -1,19 +1,19 @@
 import { defineStorybookActionsAndVModels } from "@sit-onyx/storybook-utils";
 import type { Meta, StoryObj } from "@storybook/vue3";
-import OnyxSeparator from "./OnyxSeparator.vue";
+import OnyxNavSeparator from "./OnyxNavSeparator.vue";
 
 /**
  * Separator for the context area of the `OnyxNavBar`.
  */
-const meta: Meta<typeof OnyxSeparator> = {
+const meta: Meta<typeof OnyxNavSeparator> = {
   title: "support/Separator",
   ...defineStorybookActionsAndVModels({
-    component: OnyxSeparator,
+    component: OnyxNavSeparator,
     events: [],
   }),
 };
 
 export default meta;
-type Story = StoryObj<typeof OnyxSeparator>;
+type Story = StoryObj<typeof OnyxNavSeparator>;
 
 export const Default = { args: {} } satisfies Story;
