+++ packages/sit-onyx/src/components/OnyxNavSeparator/OnyxNavSeparator.vue
